refactor: move initial setup code from main to conversion package
